Remove unneeded files and store the culprit in jail.txt
The following unneeded files were deleted:
- interview_info.txt
- interviews.txt
- possible_witnesses.txt

Information regarding membership status of the 4 remaining
suspects was used to narrow down the list to the culprit. Knowing
that the culprit was a member of 4 specific organisations, the
membership lists of these 4 was combined and searched for each
of the remaining suspects' names. 2 of the suspects were members
of all 4, but only one of those was male and was identified as the
culprit - Jeremy Bowers

The full command used was:

cat AAA Delta_SkyMiles Terminal_City_Library Museum_of_Bash_History
	| grep "suspect_name" | wc -l

where suspect_name was replaced by the full name of each of the
4 remaining suspects in turn. The number printed equalled the
number of times that name appeared in the concatenated list, so I
was looking for 4.

jail.txt was created and the name of the culprit, Jeremy Bowers,
was stored in it.